+++ src/v2/codegen/author.rs
@@ -1,4 +1,4 @@
-//! Author extraction for cargo manifest.
+//! Author extraction for cargo manifest (c) 2018 Cargo Developers.
 //!
 //! Obtained from [Cargo](https://github.com/rust-lang/cargo/blob/fa05862cd0c6b899b801fda0f256ac5b9bae69d9/src/cargo/ops/cargo_new.rs#L690-L750).